// proyecto-antiguo/frontend/src/types/api.ts
export interface SweetSpotItem {
  id: string;
  presupuesto: number;
  estado: string;
  cliente: string;
}